Changed login screen and allowed user creation
The changes removed password and email inputs and changing them to a single username field. This change also removed the register button, and I changed the text of the login button, but it still serves the same use. When a user already exists, the login page will be skipped and you'll land on the homepage